Backend/requirements.txt
fastapi
uvicorn
smartapi-python
pyotp
python-dotenv
pandas
logzero
